Include "remotepairing-tunnel" as a service for including a device
Added to the previous service "remotepairing".  This might be enabling rsd tunnels to devices without developer images mounted, but it seems to work correctly (even if the tunnel isn't super useful to us at the moment), and is preferable to failure.

(pmd3 also creates tunnels to these devices, and we will detect missing dev images other ways)